# backend/postman/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MeeSell API",
    "description": "AI-powered operating system for Meesho suppliers",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/products": {
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Create a product (Feature 3 Fast Catalog Form)",
        "description": "§10.B.1 — POST /products.\n\nPlan-guard ``product_count`` (100 active cap) is enforced INSIDE the\nservice (step 1 of the §10.B.1 flow), not in this handler.\n\nStatus codes: 201, 400, 401, 402 (rate-limit or plan-cap),\n404 (catalog / category not found), 422 (profile incomplete).",
        "operationId": "create_product_api_v1_products_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Paginated product listing (Feature 8 — Tracking Dashboard)",
        "description": "§13.B.1 — GET /api/v1/products.\n\nComposes the seller's tracking-dashboard view by aggregating:\n\n* :func:`catalog.service.list_products` — paginated active products\n  (most-recently-updated first per §10.D).\n* :func:`customer.service.get_onboarding_completeness` — profile\n  completeness badge data.\n\nEmpty inventory returns 200 with ``products=[]`` + ``total=0`` (NOT 404\n— empty inventory is a valid state for first-time sellers).\n\nPer §13.A.1 amendment (2026-06-07): ``status_filter`` and ``search``\ndeferred to V1.5. V1 ships ``page`` + ``limit`` only.\n\nStatus codes: 200, 400 (``validation.dashboard.invalid_pagination`` —\nraised by Pydantic ``Query`` validators on out-of-range inputs), 401.\n\nNo audit event (read-only per §13.B + §4.G).\nNo plan_guard (§13.I lock — dashboard is one of 3 plan_guard-excluded\nmodules alongside customer + pricing).",
        "operationId": "list_products_api_v1_products_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "patch": {
        "tags": [
          "catalog"
        ],
        "summary": "Update product fields (autosave + manual save)",
        "description": "§10.B.2 — PATCH /products/{id}.\n\n``X-Autosave: true`` header signals autosave (vs manual \"Save\").\nWhen present, the service additionally upserts a ``product_drafts``\nrow per `MVP_ARCH §11.6`.\n\nStatus codes: 200, 400, 401, 404, 422.",
        "operationId": "patch_product_api_v1_products__id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          },
          {
            "name": "x-autosave",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autosave"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get a single product — returns category_id for CatalogForm reload (GAP-1 fix)",
        "description": "§10.B.7 — GET /products/{id}.\n\nNo audit event (read-only per MVP_ARCH §11.3).\nNo feature flag — GAP-1 fix is unconditional.\n\nStatus codes: 200, 401, 404.",
        "operationId": "get_product_api_v1_products__id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "catalog"
        ],
        "summary": "Soft-delete a product",
        "description": "§10.B.5 — DELETE /products/{id}.\n\nStatus codes: 204, 401, 404.",
        "operationId": "delete_product_api_v1_products__id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/autofill": {
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "AI Auto-fill — Gemini-driven field suggestions",
        "description": "§10.B.3 — POST /products/{id}/autofill.\n\nReturns 200 with ``fallback_offered=true`` when AI is unavailable\n(budget exhaustion, guardrail retry exhaustion).  503 is reserved\nfor unrecoverable Gemini SDK exhaustion beyond §6A's fallback\ncontract (raised inside the service as\n:class:`exceptions.AutofillFailedError`).\n\nPlan-guard ``ai_autofill_hourly`` (50/h/user) is enforced INSIDE\nthe service (step 2 of the §10.B.3 flow).\n\nFeature flag: returns 404 when ``FEATURE_AI_AUTOFILL_ENABLED=false``\nper Master Plan §3.2 + ai-autofill FEATURE_PLAN.md D2.  Reading\n``settings`` at request-time (not import-time) so tests can\nmonkeypatch the attribute and the guard reflects the override.\n\nStatus codes: 200 (success OR graceful fallback), 400, 401, 402,\n404, 422.",
        "operationId": "autofill_product_api_v1_products__id__autofill_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutofillRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutofillResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/draft": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Draft recovery — most recent autosave snapshot",
        "description": "§10.B.6 — GET /products/{id}/draft.\n\nReturns 200 with the snapshot when a draft exists; 204 with no body\nwhen the product has never been autosaved.\n\nNo audit event (read-only).\n\nStatus codes: 200, 204, 401, 404.",
        "operationId": "get_product_draft_api_v1_products__id__draft_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDraftResponse"
                }
              }
            }
          },
          "204": {
            "description": "No draft snapshot exists for this product."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AutofillRequest": {
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Description"
          },
          "fields_to_fill": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields To Fill"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "description"
        ],
        "title": "AutofillRequest",
        "description": "POST /api/v1/products/{id}/autofill body per §10.B.3.\n\n* ``description`` 1..2000 chars per §10.E.\n* ``fields_to_fill`` optional allowlist of canonical names — default\n  ``None`` lets the AI fill all empty compulsory fields it has\n  confidence in."
      },
      "AutofillResponse": {
        "properties": {
          "suggestions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AutofillSuggestion"
            },
            "type": "object",
            "title": "Suggestions"
          },
          "applied": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Applied"
          },
          "fallback_offered": {
            "type": "boolean",
            "title": "Fallback Offered",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "suggestions",
          "applied"
        ],
        "title": "AutofillResponse",
        "description": "200 response for §10.B.3.\n\nAlways HTTP 200 — even when ``fallback_offered=True`` (§9.B.1 +\n§6A.F precedent: budget exhaustion does NOT 503)."
      },
...
      "CreateProductRequest": {
        "properties": {
          "catalog_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Id"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "category_id"
        ],
        "title": "CreateProductRequest",
        "description": "POST /api/v1/products body per §10.B.1.\n\n* ``catalog_id`` ``None`` → service creates a default-named catalog\n  (``\"{seller_phone_last4}-Drafts-{YYYYMMDD-HHMM}\"``).  Non-null →\n  product is added to the existing catalog (404 if not owned).\n* ``name`` is the seller-visible product name; ``None`` lets the\n  wizard default to \"Untitled product\"."
      },
      "DashboardResponse": {
        "properties": {
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductListItem"
            },
            "type": "array",
            "title": "Products"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "onboarding_completeness": {
            "$ref": "#/components/schemas/ProfileCompletenessSummary"
          }
        },
        "type": "object",
        "required": [
          "products",
          "total",
          "page",
          "limit",
          "onboarding_completeness"
        ],
        "title": "DashboardResponse",
        "description": "Wire shape for ``GET /api/v1/products`` (200 OK)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatchProductRequest": {
        "properties": {
          "fields": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "draft",
                  "ready"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PatchProductRequest",
        "description": "PATCH /api/v1/products/{id} body per §10.B.2.\n\n* ``fields`` is a partial JSON patch — keys are canonical field names\n  per ``templates.schema_jsonb.fields[*].canonical_name`` (§5A.C\n  regex ``[a-z][a-z0-9_]*``), values are primitive payloads per the\n  field's ``primitive`` (§5A.D).\n* ``status`` is the optional state transition: omit to leave\n  unchanged; ``\"ready\"`` triggers full-schema completeness validation;\n  ``\"draft\"`` is the explicit revert.\n\nAt least one of ``fields`` / ``status`` MUST be present — the\nvalidator below rejects empty bodies with 400."
      },
...
      "ProductDraftResponse": {
        "properties": {
          "fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Fields"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated"
          },
          "autosave_count": {
            "type": "integer",
            "title": "Autosave Count"
          }
        },
        "type": "object",
        "required": [
          "fields",
          "last_updated",
          "autosave_count"
        ],
        "title": "ProductDraftResponse",
        "description": "200 response for §10.B.6 (204 when no draft — empty body).\n\nDECISION FLAG §10-CATALOG-D1 (applied):\n    ``last_updated`` ← ``product_drafts.saved_at``.\n    ``fields`` ← ``draft_jsonb[\"fields\"]`` (wrapper).\n    ``autosave_count`` ← ``draft_jsonb[\"autosave_count\"]``.\n\nLegacy rows without the wrapper key coerce to\n``autosave_count=1`` and ``fields=<draft_jsonb>``."
      },
...
      "ProductResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "catalog_id": {
            "type": "string",
            "format": "uuid",
            "title": "Catalog Id"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "ready"
            ],
            "title": "Status"
          },
          "fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Fields"
          },
          "ai_suggestions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Suggestions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "catalog_id",
          "category_id",
          "name",
          "status",
          "fields",
          "created_at",
          "updated_at"
        ],
        "title": "ProductResponse",
        "description": "201 / 200 response across §10.B.1 + §10.B.2 + §10.B.3.\n\n``fields`` mirrors ``products.fields_jsonb``;\n``ai_suggestions`` mirrors ``products.ai_suggestions_jsonb``."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/otp/verify"
          }
        }
      }
    }
  }
}